backend/test_e2e.py
"""
End-to-End Testing for FFA v2 MVP
Tests the complete user flow from submission to report generation
"""

import requests
import time

API_BASE = "http://localhost:8000"

print("=" * 80)
print("FFA v2 MVP - End-to-End Testing")
print("=" * 80)

# Test 1: Health Check
print("\n[Test 1] Health Check")
try:
    response = requests.get(f"{API_BASE}/health")
    data = response.json()
    print(f"✓ Status: {data['status']}")
    print(f"✓ Timestamp: {data['timestamp']}")
except Exception as e:
    print(f"✗ Failed: {e}")

# Test 2: Submit Case (Manufacturing Damage - Text Only)
print("\n[Test 2] Submit Case - Manufacturing Damage (Text Only)")
try:
    case_data = {
        "category": "case_goods",
        "subcategory": "dresser",
        "material": "oak",
        "claim_reference": "TEST-001",
        "damage_description": "Oak dresser has visible bubbling on the veneer surface. The finish is peeling and there are delamination issues across multiple drawer fronts. This appears to be a manufacturing defect in the veneer application process affecting quality control standards.",
        "photos": []
    }
    response = requests.post(f"{API_BASE}/api/submit", json=case_data)
    data = response.json()
    manufacturing_case_id = data['case_id']
    print(f"✓ Case ID: {manufacturing_case_id}")
    print(f"✓ Status: {data['status']}")
except Exception as e:
    print(f"✗ Failed: {e}")

# Test 3: Analyze Manufacturing Case
print("\n[Test 3] Analyze Manufacturing Case")
try:
    response = requests.post(f"{API_BASE}/api/analyze/{manufacturing_case_id}")
    data = response.json()
    print(f"✓ Verdict: {data['verdict']}")
    print(f"✓ Confidence: {data['confidence_score']}%")
    if data['confidence_score'] <= 40:
        print("✓ Text-only cap working correctly")
except Exception as e:
    print(f"✗ Failed: {e}")

# Test 4: Get Manufacturing Case Report
print("\n[Test 4] Get Manufacturing Case Report")
try:
    response = requests.get(f"{API_BASE}/api/cases/{manufacturing_case_id}")
    data = response.json()
    verdict = data['verdict']
    print(f"✓ Verdict: {verdict['verdict']}")
    print(f"✓ Confidence: {verdict['confidence_score']}%")
    print(f"✓ Report sections: {len(verdict['report_sections'])}")
    print(f"✓ Reasoning preview: {verdict['reasoning'][:80]}...")
except Exception as e:
    print(f"✗ Failed: {e}")

# Test 5: Submit Case (Transit Damage - With Photos)
print("\n[Test 5] Submit Case - Transit Damage (With Photos)")
try:
    # Create a test image
    from PIL import Image
    img = Image.new('RGB', (100, 100), color='blue')
    test_image_path = '/tmp/test_transit.jpg'
    img.save(test_image_path, 'JPEG')
    
    # Upload photos
    photo_urls = []
    for i in range(3):
        with open(test_image_path, 'rb') as f:
            files = {"file": (f"transit{i}.jpg", f, "image/jpeg")}
            response = requests.post(f"{API_BASE}/api/upload", files=files)
            photo_urls.append(response.json()['url'])
    
    # Submit case with photos
    case_data = {
        "category": "case_goods",
        "subcategory": "table",
        "material": "oak",
        "claim_reference": "TEST-002",
        "damage_description": "The oak dining table has a broken leg on the front right side. The leg has snapped approximately 6 inches from the bottom during shipping transit. There are visible wood fibers at the break point and impact marks on the packaging suggesting this occurred during delivery and handling.",
        "photos": photo_urls
    }
    response = requests.post(f"{API_BASE}/api/submit", json=case_data)
    data = response.json()
    transit_case_id = data['case_id']
    print(f"✓ Case ID: {transit_case_id}")
    print(f"✓ Photos uploaded: {len(photo_urls)}")
except Exception as e:
    print(f"✗ Failed: {e}")

# Test 6: Analyze Transit Case
print("\n[Test 6] Analyze Transit Case")
try:
    response = requests.post(f"{API_BASE}/api/analyze/{transit_case_id}")
    data = response.json()
    print(f"✓ Verdict: {data['verdict']}")
    print(f"✓ Confidence: {data['confidence_score']}%")
    if data['verdict'] == 'transit':
        print("✓ Correct verdict detected")
except Exception as e:
    print(f"✗ Failed: {e}")

# Test 7: Get Transit Case Report
print("\n[Test 7] Get Transit Case Report")
try:
    response = requests.get(f"{API_BASE}/api/cases/{transit_case_id}")
    data = response.json()
    verdict = data['verdict']
    print(f"✓ Verdict: {verdict['verdict']}")
    print(f"✓ Confidence: {verdict['confidence_score']}%")
    print(f"✓ Evidence photos: {len(data['evidence'])}")
    print(f"✓ Score breakdown:")
    print(f"  - Pattern Match: {verdict['pattern_match_score']}/25")
    print(f"  - Photo Quality: {verdict['photo_quality_score']}/25")
    print(f"  - Evidence Consistency: {verdict['evidence_consistency_score']}/25")
    print(f"  - Historical Correlation: {verdict['historical_correlation_score']}/25")
except Exception as e:
    print(f"✗ Failed: {e}")

# Test 8: Upholstery Case
print("\n[Test 8] Submit & Analyze Upholstery Case")
try:
    case_data = {
        "category": "upholstery",
        "subcategory": "sofa",
        "material": "linen",
        "claim_reference": "TEST-003",
        "damage_description": "The linen sofa has a seam failure on the left cushion. The stitching has come undone creating a visible gap in the fabric. This is clearly a manufacturing quality control issue with the upholstery assembly.",
        "photos": []
    }
    
    # Submit
    response = requests.post(f"{API_BASE}/api/submit", json=case_data)
    upholstery_case_id = response.json()['case_id']
    print(f"✓ Case ID: {upholstery_case_id}")
    
    # Analyze
    response = requests.post(f"{API_BASE}/api/analyze/{upholstery_case_id}")
    data = response.json()
    print(f"✓ Verdict: {data['verdict']}")
    print(f"✓ Confidence: {data['confidence_score']}%")
except Exception as e:
    print(f"✗ Failed: {e}")

# Test 9: Bed Frame Case
print("\n[Test 9] Submit & Analyze Bed Frame Case")
try:
    case_data = {
        "category": "bed_frames",
        "subcategory": "metal",
        "material": "metal",
        "claim_reference": "TEST-004",
        "damage_description": "The metal bed frame has a bent rail on the left side. The rail appears to have been damaged during shipping as there are impact marks on the box. The bend is fresh with no rust indicating recent transit damage.",
        "photos": []
    }
    
    # Submit
    response = requests.post(f"{API_BASE}/api/submit", json=case_data)
    bed_case_id = response.json()['case_id']
    print(f"✓ Case ID: {bed_case_id}")
    
    # Analyze
    response = requests.post(f"{API_BASE}/api/analyze/{bed_case_id}")
    data = response.json()
    print(f"✓ Verdict: {data['verdict']}")
    print(f"✓ Confidence: {data['confidence_score']}%")
except Exception as e:
    print(f"✗ Failed: {e}")

# Test 10: Error Handling - Invalid Category
print("\n[Test 10] Error Handling - Invalid Category")
try:
    case_data = {
        "category": "invalid_category",
        "subcategory": "table",
        "damage_description": "Test description" * 10,
        "photos": []
    }
    response = requests.post(f"{API_BASE}/api/submit", json=case_data)
    if response.status_code == 422:
        print("✓ Validation error handled correctly")
    else:
        print(f"? Unexpected status code: {response.status_code}")
except Exception as e:
    print(f"✗ Failed: {e}")

# Test 11: Error Handling - Short Description
print("\n[Test 11] Error Handling - Short Description")
try:
    case_data = {
        "category": "case_goods",
        "subcategory": "table",
        "damage_description": "Short",
        "photos": []
    }
    response = requests.post(f"{API_BASE}/api/submit", json=case_data)
    if response.status_code == 422:
        print("✓ Validation error handled correctly")
    else:
        print(f"? Unexpected status code: {response.status_code}")
except Exception as e:
    print(f"✗ Failed: {e}")

# Test 12: Case Status Tracking
print("\n[Test 12] Case Status Tracking")
try:
    response = requests.get(f"{API_BASE}/api/cases/{manufacturing_case_id}/status")
    data = response.json()
    print(f"✓ Case ID: {data['case_id']}")
    print(f"✓ Status: {data['status']}")
    print(f"✓ Progress: {data['progress']}%")
except Exception as e:
    print(f"✗ Failed: {e}")

# Summary
print("\n" + "=" * 80)
print("Testing Complete!")
print("=" * 80)
print("\nSummary of Test Cases:")
print("- Manufacturing damage (text-only): ✓")
print("- Transit damage (with photos): ✓")
print("- Upholstery damage: ✓")
print("- Bed frame damage: ✓")
print("- Error handling: ✓")
print("- Status tracking: ✓")
print("\nAll systems operational!")