backend/requirements.txt
# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Data Validation
pydantic==2.11.1
pydantic-settings==2.5.2
anyio>=4.7.0

# Image Processing
Pillow==10.1.0

# CORS
python-dotenv==1.0.0

# Logging
python-json-logger==2.0.7

# Testing (for later)
pytest==7.4.3
pytest-asyncio==0.21.1